fix reloading bugs

+++ iavlx/load.go
@@ -21,11 +21,14 @@ func (ts *TreeStore) load() error {
 		if !de.IsDir() {
 			continue
 		}
-		dir := filepath.Join(ts.dir, de.Name())
-		startVersion, compactedAt, valid := ParseChangesetDirName(dir)
+
+		dirName := de.Name()
+		startVersion, compactedAt, valid := ParseChangesetDirName(dirName)
 		if !valid {
 			continue
 		}
+
+		dir := filepath.Join(ts.dir, dirName)
 		if _, found := dirMap.Get(startVersion); !found {
 			dirMap.Set(startVersion, &btree.Map[uint64, string]{})
 		}
@@ -40,13 +43,10 @@ func (ts *TreeStore) load() error {
 			return nil
 		}
 
-		_, dirName, ok := compactionMap.PopMax()
-		if !ok {
-			return fmt.Errorf("internal error: no changeset entries for start version %d", startVersion)
-		}
+		// startVersion should be equal to stagedVersion
 
 		if startVersion < uint64(ts.stagedVersion) {
-			ts.logger.Warn("found undeleted compaction", "startVersion", startVersion, "stagedVersion", ts.stagedVersion, "dir", dirName)
+			ts.logger.Warn("found undeleted changeset that was already compacted", "startVersion", startVersion, "stagedVersion", ts.stagedVersion)
 			// TODO delete undeleted compactions
 			continue
 		}
@@ -55,30 +55,57 @@ func (ts *TreeStore) load() error {
 			return fmt.Errorf("missing changeset for staged version %d", ts.stagedVersion)
 		}
 
-		ready, err := IsChangesetReady(dirName)
-		if err != nil {
-			return fmt.Errorf("failed to check if changeset %s is ready: %w", dirName, err)
-		}
-
-		if !ready {
-			ts.logger.Warn("found incomplete compaction, deleting", "dir", dirName)
-			err := os.RemoveAll(dirName)
-			if err != nil {
-				ts.logger.Error("failed to remove incomplete compaction", "dir", dirName, "error", err)
+		for {
+			_, dirName, ok := compactionMap.PopMax()
+			if !ok {
+				return fmt.Errorf("internal error: no changeset entries for start version %d", startVersion)
 			}
-			continue
+
+			ready, err := IsChangesetReady(dirName)
+			if err != nil {
+				return fmt.Errorf("failed to check if changeset %s is ready: %w", dirName, err)
+			}
+
+			if !ready {
+				ts.logger.Warn("found incomplete compaction, deleting", "dir", dirName)
+				err := os.RemoveAll(dirName)
+				if err != nil {
+					ts.logger.Error("failed to remove incomplete compaction", "dir", dirName, "error", err)
+				}
+				continue
+			}
+
+			ts.logger.Debug("loading changeset", "startVersion", startVersion, "dir", dirName)
+
+			cs, err := OpenChangeset(ts, dirName)
+			if err != nil {
+				return fmt.Errorf("failed to open changeset in %s: %w", dirName, err)
+			}
+
+			realStartVersion := cs.info.StartVersion
+			if uint64(realStartVersion) != startVersion {
+				if realStartVersion == 0 {
+					if dirMap.Len() != 0 {
+						return fmt.Errorf("found incomplete changeset %s, but there are later changesets present", dirName)
+					}
+					ts.logger.Debug("found final incomplete changeset, deleting", "dir", dirName)
+					err := os.RemoveAll(dirName)
+					if err != nil {
+						return fmt.Errorf("failed to remove incomplete changeset %s: %w", dirName, err)
+					}
+					break
+				} else {
+					return fmt.Errorf("changeset in %s has mismatched start version %d (expected %d)", dirName, realStartVersion, startVersion)
+				}
+			}
+
+			ce := &changesetEntry{}
+			ce.changeset.Store(cs)
+			ts.changesets.Set(uint32(startVersion), ce)
+
+			ts.savedVersion.Store(cs.info.EndVersion)
+			ts.stagedVersion = cs.info.EndVersion + 1
+			break
 		}
-
-		cs, err := OpenChangeset(ts, dirName)
-		if err != nil {
-			return fmt.Errorf("failed to open changeset in %s: %w", dirName, err)
-		}
-
-		ce := &changesetEntry{}
-		ce.changeset.Store(cs)
-		ts.changesets.Set(uint32(startVersion), ce)
-
-		ts.savedVersion.Store(cs.info.EndVersion)
-		ts.stagedVersion = cs.info.EndVersion + 1
 	}
 }
